+++ src/types/common/button.type.ts
@@ -1,9 +1,10 @@
 import React from 'react';
 
 export interface ButtonPropsType {
+  value?: string;
   padding?: string;
   width?: string;
-  option: 'UNFILLED' | 'FILLED';
-  onClick: React.MouseEventHandler<HTMLButtonElement>;
-  element: React.ReactNode;
+  option: 'UNFILLED' | 'FILLED' | 'SCARCE_FILLED';
+  onClick?: React.MouseEventHandler<HTMLButtonElement>;
+  element?: React.ReactNode;
 }
